Add readmes of packages to website
Also:

* Fix prev/next sorting
* Add support for markdown as server/client components
