+++ docs/_component/nav.server.js
@@ -1,57 +1,19 @@
 import React from 'react'
 import apStyleTitleCase from 'ap-style-title-case'
-import dlv from 'dlv'
+import {sortItems} from './sort.js'
 
-const collator = new Intl.Collator('en').compare
 const dateTimeFormat = new Intl.DateTimeFormat('en')
 
 export const NavGroup = (props) => {
   const {items, className, sort = 'navSortSelf,meta.title', ...rest} = props
-  const fields = sort.split(',').map((d) => {
-    const [field, order = 'asc'] = d.split(':')
-
-    if (order !== 'asc' && order !== 'desc') {
-      throw new Error('Cannot order as `' + order + '`')
-    }
-
-    return [field, order]
-  })
-
-  const list = [...items].sort(sortFn)
 
   return (
     <ol {...{className}}>
-      {list.map((d) => (
+      {sortItems(items, sort).map((d) => (
         <NavItem key={d.name} {...rest} item={d} />
       ))}
     </ol>
   )
-
-  function sortFn(left, right) {
-    let index = -1
-
-    while (++index < fields.length) {
-      const [field, order] = fields[index]
-      let a = dlv(left.data, field)
-      let b = dlv(right.data, field)
-
-      if (a && typeof a === 'object' && 'valueOf' in a) a = a.valueOf()
-      if (b && typeof b === 'object' && 'valueOf' in b) b = b.valueOf()
-
-      const score =
-        typeof a === 'number' || typeof b === 'number'
-          ? a === null || a === undefined
-            ? 1
-            : b === null || b === undefined
-            ? -1
-            : a - b
-          : collator(a, b)
-      const result = order === 'asc' ? score : -score
-      if (result) return result
-    }
-
-    return 0
-  }
 }
 
 export const NavItem = (props) => {
