+++ docs/packages/loader.server.md
@@ -1,0 +1,1 @@
+../../packages/loader/readme.md
