+++ docs/packages/mdx.server.md
@@ -1,0 +1,1 @@
+../../packages/mdx/readme.md
