+++ docs/packages/node-loader.server.md
@@ -1,0 +1,1 @@
+../../packages/node-loader/readme.md
