+++ docs/packages/preact.server.md
@@ -1,0 +1,1 @@
+../../packages/preact/readme.md
