+++ docs/packages/react.server.md
@@ -1,0 +1,1 @@
+../../packages/react/readme.md
