+++ docs/packages/register.server.md
@@ -1,0 +1,1 @@
+../../packages/register/readme.md
