+++ docs/packages/remark-mdx.server.md
@@ -1,0 +1,1 @@
+../../packages/remark-mdx/readme.md
