+++ docs/packages/rollup.server.md
@@ -1,0 +1,1 @@
+../../packages/rollup/readme.md
