+++ docs/packages/vue.server.md
@@ -1,0 +1,1 @@
+../../packages/vue/readme.md
