+++ packages/esbuild/readme.md
@@ -212,7 +212,7 @@ abide by its terms.
 
 [coc]: https://github.com/mdx-js/.github/blob/master/code-of-conduct.md
 
-[mit]: license
+[mit]: https://github.com/mdx-js/mdx/blob/main/packages/esbuild/license
 
 [author]: https://wooorm.com
