+++ packages/loader/readme.md
@@ -192,7 +192,7 @@ abide by its terms.
 
 [coc]: https://github.com/mdx-js/.github/blob/master/code-of-conduct.md
 
-[mit]: license
+[mit]: https://github.com/mdx-js/mdx/blob/main/packages/loader/license
 
 [vercel]: https://vercel.com
