+++ packages/mdx/readme.md
@@ -1,5 +1,3 @@
-<!-- lint disable no-html -->
-
 # `@mdx-js/mdx`
 
 [![Build][build-badge]][build]
@@ -848,7 +846,7 @@ abide by its terms.
 
 [coc]: https://github.com/mdx-js/.github/blob/master/code-of-conduct.md
 
-[mit]: license
+[mit]: https://github.com/mdx-js/mdx/blob/main/packages/mdx/license
 
 [compositor]: https://compositor.io
