+++ packages/preact/readme.md
@@ -213,6 +213,6 @@ abide by its terms.
 
 [coc]: https://github.com/mdx-js/.github/blob/master/code-of-conduct.md
 
-[mit]: license
+[mit]: https://github.com/mdx-js/mdx/blob/main/packages/preact/license
 
 [vercel]: https://vercel.com
