+++ packages/react/readme.md
@@ -213,7 +213,7 @@ abide by its terms.
 
 [coc]: https://github.com/mdx-js/.github/blob/master/code-of-conduct.md
 
-[mit]: license
+[mit]: https://github.com/mdx-js/mdx/blob/main/packages/react/license
 
 [compositor]: https://compositor.io
