+++ packages/register/readme.md
@@ -168,6 +168,6 @@ abide by its terms.
 
 [coc]: https://github.com/mdx-js/.github/blob/master/code-of-conduct.md
 
-[mit]: license
+[mit]: https://github.com/mdx-js/mdx/blob/main/packages/register/license
 
 [author]: https://wooorm.com
