+++ packages/rollup/readme.md
@@ -189,7 +189,7 @@ abide by its terms.
 
 [coc]: https://github.com/mdx-js/.github/blob/master/code-of-conduct.md
 
-[mit]: license
+[mit]: https://github.com/mdx-js/mdx/blob/main/packages/rollup/license
 
 [author]: https://wooorm.com
