+++ packages/vue/readme.md
@@ -182,6 +182,6 @@ abide by its terms.
 
 [coc]: https://github.com/mdx-js/.github/blob/master/code-of-conduct.md
 
-[mit]: license
+[mit]: https://github.com/mdx-js/mdx/blob/main/packages/vue/license
 
 [vercel]: https://vercel.com
