+++ website/generate.server.js
@@ -18,7 +18,7 @@ main().catch((error) => {
 
 async function main() {
   const files = (
-    await globby(['**/*.server.mdx', '!_component/*'], {
+    await globby(['**/*.server.{md,mdx}', '!_component/*'], {
       cwd: fileURLToPath(config.input)
     })
   ).map((d) => new URL(d, config.input))
@@ -31,7 +31,7 @@ async function main() {
     files.map((url) => async () => {
       const name = url.href
         .slice(config.input.href.length - 1)
-        .replace(/\.server\.mdx$/, '/')
+        .replace(/\.server\.mdx?$/, '/')
         .replace(/\/index\/$/, '/')
       const nljsonUrl = new URL('.' + name + 'index.nljson', config.output)
       const jsonUrl = new URL('.' + name + 'index.json', config.output)
